+++ libs/frontend/presentation/codelab-ui/src/index.ts
@@ -3,4 +3,5 @@ export * from './components'
 export * from './core'
 export * from './form'
 export * from './layout'
+export * from './provider'
 export * from './views'
