+++ libs/frontend/presentation/codelab-ui/src/provider/DynamicProviderTree.tsx
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+export const DynamicProvidersTree = dynamic(() => import('./ProviderTree'), {
+  ssr: false,
+})
